# birthday-invite-atualizado/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Birthday invite RSVP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /guests:
    get:
      operationId: listGuests
      tags: [guests]
      summary: List guests (public)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 6
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of guests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestListResponse"
    post:
      operationId: createGuest
      tags: [guests]
      summary: Confirm presence (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGuestBody"
      responses:
        "201":
          description: Guest created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /guests/{id}:
    get:
      operationId: getGuest
      tags: [guests]
      summary: Get a guest by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Guest found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateGuest
      tags: [guests]
      summary: Update a guest (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGuestBody"
      responses:
        "200":
          description: Guest updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteGuest
      tags: [guests]
      summary: Delete a guest (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get event statistics (admin)
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/confirmed-guests:
    get:
      operationId: listConfirmedGuests
      tags: [admin]
      summary: List confirmed guests with phone (for mass reminder)
      responses:
        "200":
          description: Confirmed guests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/guest-audit:
    get:
      operationId: listGuestAudit
      tags: [admin]
      summary: List guest audit log
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: guestId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestAuditListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/photos:
    post:
      operationId: createPhoto
      tags: [photos]
      summary: Add a photo to the gallery (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePhotoBody"
      responses:
        "201":
          description: Photo created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/photos/{id}:
    patch:
      operationId: updatePhoto
      tags: [photos]
      summary: Update a photo (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePhotoBody"
      responses:
        "200":
          description: Photo updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePhoto
      tags: [photos]
      summary: Delete a photo (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Guest:
      type: object
      properties:
        id:
          type: integer
        parentName:
          type: string
        childName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        adultsCount:
          type: integer
        childrenCount:
          type: integer
        status:
          type: string
          enum: [confirmed, maybe, declined]
        notes:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - parentName
        - adultsCount
        - childrenCount
        - status
        - createdAt
        - updatedAt
    CreateGuestBody:
      type: object
      properties:
        parentName:
          type: string
          minLength: 2
        childName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        adultsCount:
          type: integer
          minimum: 1
        childrenCount:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [confirmed, maybe, declined]
        notes:
          type: ["string", "null"]
      required:
        - parentName
        - adultsCount
        - childrenCount
        - status
    UpdateGuestBody:
      type: object
      properties:
        parentName:
          type: string
          minLength: 2
        childName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        adultsCount:
          type: integer
          minimum: 1
        childrenCount:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [confirmed, maybe, declined]
        notes:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
    GuestListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Guest"
        page:
          type: integer
        limit:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
      required:
        - items
        - page
        - limit
        - totalItems
        - totalPages
    Photo:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        caption:
          type: ["string", "null"]
        displayOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - url
        - displayOrder
        - createdAt
    CreatePhotoBody:
      type: object
      properties:
        url:
          type: string
          minLength: 1
        caption:
          type: ["string", "null"]
        displayOrder:
          type: integer
      required:
        - url
    UpdatePhotoBody:
      type: object
      properties:
        caption:
          type: ["string", "null"]
        displayOrder:
          type: integer
    GuestAudit:
      type: object
      properties:
        id:
          type: integer
        guestId:
          type: ["integer", "null"]
        guestName:
          type: string
        action:
          type: string
        previousData:
          nullable: true
        newData:
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - guestName
        - action
        - createdAt
    GuestAuditListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GuestAudit"
        totalItems:
          type: integer
        totalPages:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - items
        - totalItems
        - totalPages
        - page
        - limit
    AdminLoginBody:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    AdminLoginResponse:
      type: object
      properties:
        token:
          type: string
        message:
          type: string
      required:
        - token
        - message
    AdminStats:
      type: object
      properties:
        total:
          type: integer
        confirmed:
          type: integer
        maybe:
          type: integer
        declined:
          type: integer
        totalAdults:
          type: integer
        totalChildren:
          type: integer
      required:
        - total
        - confirmed
        - maybe
        - declined
        - totalAdults
        - totalChildren
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error