# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Birthday invite RSVP API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GuestStatus = (typeof GuestStatus)[keyof typeof GuestStatus];

export const GuestStatus = {
  confirmed: "confirmed",
  maybe: "maybe",
  declined: "declined",
} as const;

export interface Guest {
  id: number;
  parentName: string;
  /** @nullable */
  childName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  adultsCount: number;
  childrenCount: number;
  status: GuestStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateGuestBodyStatus =
  (typeof CreateGuestBodyStatus)[keyof typeof CreateGuestBodyStatus];

export const CreateGuestBodyStatus = {
  confirmed: "confirmed",
  maybe: "maybe",
  declined: "declined",
} as const;

export interface CreateGuestBody {
  /** @minLength 2 */
  parentName: string;
  /** @nullable */
  childName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @minimum 1 */
  adultsCount: number;
  /** @minimum 0 */
  childrenCount: number;
  status: CreateGuestBodyStatus;
  /** @nullable */
  notes?: string | null;
}

export type UpdateGuestBodyStatus =
  (typeof UpdateGuestBodyStatus)[keyof typeof UpdateGuestBodyStatus];

export const UpdateGuestBodyStatus = {
  confirmed: "confirmed",
  maybe: "maybe",
  declined: "declined",
} as const;

export interface UpdateGuestBody {
  /** @minLength 2 */
  parentName?: string;
  /** @nullable */
  childName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @minimum 1 */
  adultsCount?: number;
  /** @minimum 0 */
  childrenCount?: number;
  status?: UpdateGuestBodyStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  adminNotes?: string | null;
}

export interface GuestListResponse {
  items: Guest[];
  page: number;
  limit: number;
  totalItems: number;
  totalPages: number;
}

export interface Photo {
  id: number;
  url: string;
  /** @nullable */
  caption?: string | null;
  displayOrder: number;
  createdAt: string;
}

export interface CreatePhotoBody {
  /** @minLength 1 */
  url: string;
  /** @nullable */
  caption?: string | null;
  displayOrder?: number;
}

export interface UpdatePhotoBody {
  /** @nullable */
  caption?: string | null;
  displayOrder?: number;
}

export interface GuestAudit {
  id: number;
  /** @nullable */
  guestId?: number | null;
  guestName: string;
  action: string;
  previousData?: unknown | null;
  newData?: unknown | null;
  createdAt: string;
}

export interface GuestAuditListResponse {
  items: GuestAudit[];
  totalItems: number;
  totalPages: number;
  page: number;
  limit: number;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResponse {
  token: string;
  message: string;
}

export interface AdminStats {
  total: number;
  confirmed: number;
  maybe: number;
  declined: number;
  totalAdults: number;
  totalChildren: number;
}

export interface MessageResponse {
  message: string;
}

export interface EventConfig {
  id: number;
  childName: string;
  age: string;
  dateLabel: string;
  dateFull: string;
  timeLabel: string;
  location: string;
  neighborhood: string;
  tagline: string;
  inviteImageUrl?: string | null;
  heroBgFrom: string;
  heroBgVia: string;
  heroBgTo: string;
  musicUrl?: string | null;
  galleryEnabled: boolean;
  galleryTitle: string;
  theme: string;
  /** @nullable */
  spotifyPlaylistUrl?: string | null;
  /** @nullable */
  mapsUrl?: string | null;
  whatsappReminderEnabled?: boolean;
  whatsappReminderDaysBefore?: string;
  updatedAt: string;
}

export interface UpdateEventConfigBody {
  /** @minLength 1 */
  childName?: string;
  /** @minLength 1 */
  age?: string;
  /** @minLength 1 */
  dateLabel?: string;
  /** @minLength 1 */
  dateFull?: string;
  /** @minLength 1 */
  timeLabel?: string;
  /** @minLength 1 */
  location?: string;
  /** @minLength 1 */
  neighborhood?: string;
  /** @minLength 1 */
  tagline?: string;
  inviteImageUrl?: string | null;
  heroBgFrom?: string;
  heroBgVia?: string;
  heroBgTo?: string;
  musicUrl?: string | null;
  galleryEnabled?: boolean;
  galleryTitle?: string;
  theme?: string;
  /** @nullable */
  spotifyPlaylistUrl?: string | null;
  /** @nullable */
  mapsUrl?: string | null;
  whatsappReminderEnabled?: boolean;
  whatsappReminderDaysBefore?: string;
}

export interface SendWhatsAppBody {
  /** @minLength 1 */
  message: string;
  guestIds?: number[];
}

export interface SendWhatsAppResponse {
  sent: number;
  failed: number;
  errors: string[];
}

export interface WhatsAppStatus {
  configured: boolean;
  /** @nullable */
  phoneId: string | null;
}

export interface PublicStats {
  totalFamilies: number;
  totalPeople: number;
  totalAdults: number;
  totalChildren: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListGuestsParams = {
  page?: number;
  limit?: number;
  status?: string;
  search?: string;
};

export type ListGuestAuditParams = {
  page?: number;
  limit?: number;
  guestId?: number;
};