# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Birthday invite RSVP API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  total: number;
  confirmed: number;
  maybe: number;
  declined: number;
  totalAdults: number;
  totalChildren: number;
}