# lib/api-zod/src/generated/types/createGuestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Birthday invite RSVP API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateGuestBodyStatus } from "./createGuestBodyStatus";

export interface CreateGuestBody {
  /** @minLength 2 */
  parentName: string;
  /** @nullable */
  childName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @minimum 1 */
  adultsCount: number;
  /** @minimum 0 */
  childrenCount: number;
  status: CreateGuestBodyStatus;
  /** @nullable */
  notes?: string | null;
}