# lib/api-zod/src/generated/types/createPhotoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Birthday invite RSVP API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePhotoBody {
  /** @minLength 1 */
  url: string;
  /** @nullable */
  caption?: string | null;
  displayOrder?: number;
}