# lib/api-zod/src/generated/types/updatePhotoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Birthday invite RSVP API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdatePhotoBody {
  /** @nullable */
  caption?: string | null;
  displayOrder?: number;
}